+++ src/main/java/com/alth/backend/record/domain/Record.java
@@ -1,7 +1,7 @@
 package com.alth.backend.record.domain;
 
 
-import com.alth.backend.Alcohol.domain.Alcohol;
+import com.alth.backend.alcohol.domain.Alcohol;
 import com.alth.backend.global.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.*;
