+++ src/main/java/com/alth/backend/record/dto/request/RecordRequestDto.java
@@ -1,9 +1,7 @@
 package com.alth.backend.record.dto.request;
 
-import com.alth.backend.Alcohol.dto.request.AlcoholRequestDto;
+import com.alth.backend.alcohol.dto.request.AlcoholRequestDto;
 import com.alth.backend.record.domain.Feel;
-import com.alth.backend.record.domain.Record;
-import com.alth.backend.Alcohol.dto.response.AlcoholResponseListDto;
 import lombok.*;
 
 import java.util.ArrayList;
