+++ src/main/java/com/alth/backend/record/dto/request/RecordUpdateDto.java
@@ -1,10 +1,9 @@
 package com.alth.backend.record.dto.request;
 
-import com.alth.backend.Alcohol.dto.request.AlcoholRequestDto;
-import com.alth.backend.Alcohol.domain.Alcohol;
+import com.alth.backend.alcohol.dto.request.AlcoholRequestDto;
+import com.alth.backend.alcohol.domain.Alcohol;
 import com.alth.backend.record.domain.Feel;
-import com.alth.backend.record.domain.Record;
-import com.alth.backend.Alcohol.dto.response.AlcoholResponseListDto;
+import com.alth.backend.alcohol.dto.response.AlcoholResponseListDto;
 import lombok.*;
 
 import java.util.List;
