+++ src/main/java/com/alth/backend/record/dto/response/RecordResponseDto.java
@@ -1,11 +1,10 @@
 package com.alth.backend.record.dto.response;
 
-import com.alth.backend.Alcohol.domain.Alcohol;
+import com.alth.backend.alcohol.domain.Alcohol;
 import com.alth.backend.record.domain.Feel;
 import com.alth.backend.record.domain.Record;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
